requirements.txt
Flask
facebook-sdk
flask-cors